Refactor file data class, image upload handler.
